refactor: move insight merchant distribution types
