+++ frontend/src/insights/types/merchantDistribution.ts
@@ -1,0 +1,14 @@
+export type MerchantMarketMerchant = {
+  id: string
+  name: string
+  totalAmount: number
+  changePct: number | null
+  changeAmount: number | null
+}
+
+export type MerchantMarketTile = MerchantMarketMerchant & {
+  x: number
+  y: number
+  width: number
+  height: number
+}
